mealmuse_api/requirements.txt
Django
requests
lxml
dj-database-url
djangorestframework
Markdown
django-filter
gunicorn
psycopg2-binary
recipe_scrapers
beautifulsoup4
selenium
webdriver-manager
django-cors-headers
